+++ packages/vchart-extension/src/charts/ranking-list/ranking-list-transformer.ts
@@ -591,7 +591,9 @@ export class RankingListChartSpecTransformer extends CommonChartSpecTransformer 
       type: 'moveOut',
       options: {
         direction: 'y',
+
         orient: 'negative',
+        excludeChannels: ['x'],
         point: (datum: Datum, graphic: IMarkGraphic) => {
           const channelAttr = graphic.getGraphicAttribute('y');
           const barSpace = spec.height / (spec.pageSize + 1);
@@ -612,7 +614,7 @@ export class RankingListChartSpecTransformer extends CommonChartSpecTransformer 
       options: {
         direction: 'y',
         orient: 'negative',
-        excludeChannels: ['y'],
+        excludeChannels: ['x'],
         point: (datum: Datum, graphic: IMarkGraphic) => {
           const channelAttr = graphic.getGraphicAttribute('y');
           return { y: channelAttr + (spec.height / (spec.pageSize + 1)) * Math.min(spec.scrollSize, spec.pageSize) };