+++ packages/vchart-types/types/mark/group.d.ts
@@ -1,7 +1,7 @@
 import type { Maybe } from '../typings';
 import type { IGroupMarkSpec } from '../typings/visual';
 import { BaseMark } from './base/base-mark';
-import type { AnimationStateValues, IGroupMark, IMark, IMarkGraphic, MarkType } from './interface';
+import { DiffState, type IGroupMark, type IMark, type IMarkGraphic, type MarkType } from './interface';
 import { MarkTypeEnum } from './interface/type';
 import { type IMarkCompileOption } from '../compile/mark';
 import type { IGroup, IGroupGraphicAttribute } from '@visactor/vrender-core';
@@ -10,6 +10,7 @@ export declare class GroupMark extends BaseMark<IGroupMarkSpec> implements IGrou
     readonly type = MarkTypeEnum.group;
     protected _marks: IMark[];
     getMarks(): IMark[];
+    protected _diffState: DiffState;
     protected _product: Maybe<IGroup>;
     getProduct: () => Maybe<IGroup>;
     protected isMarkExist(mark: IMark): boolean;
@@ -23,8 +24,8 @@ export declare class GroupMark extends BaseMark<IGroupMarkSpec> implements IGrou
     protected _getAttrsFromConfig(attrs?: IGroupGraphicAttribute): IGroupGraphicAttribute;
     getGraphics(): IMarkGraphic[];
     renderInner(): void;
+    clearExitGraphics(): void;
     render(): void;
-    updateAnimationState(callback: (g: IMarkGraphic) => AnimationStateValues): void;
     release(): void;
 }
 export declare const registerGroupMark: () => void;