+++ packages/vchart/src/component/brush/brush.ts
@@ -297,6 +297,7 @@ export class Brush<T extends IBrushSpec = IBrushSpec> extends BaseComponent<T> i
     brush.addEventListener(BrushEvent.drawEnd, (e: any) => {
       this._setRegionMarkPickable(region, true);
       const { operateMask } = e.detail as any;
+      const { updateElementsState = true } = this._spec;
       if (this._spec?.onBrushEnd) {
         // 如果onBrushEnd返回true，则清空brush， 并抛出clear事件
         if (this._spec.onBrushEnd(e) === true) {
@@ -309,7 +310,7 @@ export class Brush<T extends IBrushSpec = IBrushSpec> extends BaseComponent<T> i
         }
       } else {
         const inBrushData = this._extendDataInBrush(this._inBrushElementsMap);
-        if (!this._spec.zoomWhenEmpty && inBrushData.length > 0) {
+        if ((!this._spec.zoomWhenEmpty && inBrushData.length > 0) || !updateElementsState) {
           this._setAxisAndDataZoom(operateMask, region);
         }
         this._emitEvent(ChartEvent.brushEnd, region);
@@ -319,8 +320,9 @@ export class Brush<T extends IBrushSpec = IBrushSpec> extends BaseComponent<T> i
     brush.addEventListener(BrushEvent.moveEnd, (e: any) => {
       this._setRegionMarkPickable(region, true);
       const { operateMask } = e.detail as any;
+      const { updateElementsState = true } = this._spec;
       const inBrushData = this._extendDataInBrush(this._inBrushElementsMap);
-      if (!this._spec.zoomWhenEmpty && inBrushData.length > 0) {
+      if ((!this._spec.zoomWhenEmpty && inBrushData.length > 0) || updateElementsState) {
         this._setAxisAndDataZoom(operateMask, region);
       }
       this._emitEvent(ChartEvent.brushEnd, region);
@@ -370,8 +372,11 @@ export class Brush<T extends IBrushSpec = IBrushSpec> extends BaseComponent<T> i
   /*** start: event dispatch ***/
   private _handleBrushChange(region: IRegion, e: any) {
     const { operateMask } = e.detail as any;
-    this._reconfigItem(operateMask, region);
-    this._reconfigLinkedItem(operateMask, region);
+    const { updateElementsState = true } = this._spec;
+    if (updateElementsState) {
+      this._reconfigItem(operateMask, region);
+      this._reconfigLinkedItem(operateMask, region);
+    }
   }
 
   protected _extendDataInBrush(elementsMap: { [brushName: string]: { [elementKey: string]: IMarkGraphic } }) {